requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
idna==2.6
instabot==0.3.3.21
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
python-decouple==3.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.8.0
schedule==0.4.3
tqdm==4.19.4
urllib3==1.22
django-crontab==0.7.1

